backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
pydantic==1.10.9
python-dotenv==1.0.0
openai>=1.0.0
redis
gunicorn